Refactor Keyboard into Controls wrapper
